app/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
openai>=1.50.0
pydantic>=2.9.0
python-dotenv==1.0.1
jinja2==3.1.4
sse-starlette==2.1.3
aiofiles==24.1.0
